Rename: svg 이름 수정
